+++ NEWS.md
@@ -1,4 +1,15 @@
-# medapps2021 0.1.0
+# medapps2021 0.0.0.9002
+
+* Added 6 cleaned datasets:
+    + `astats_major`: dataset describing mean total MCAT exam/sub-section scores and combined/science/non-science GPAs for all 2021-2022 applicants by undergraduate major in tidy format.
+    + `mstats_major`: dataset describing above information for all 2021-2022 matriculants by undergraduate major in tidy format.
+    + `astats_race`: dataset describing above information for all 2021-2022 applicants by race and ethnicity in tidy format.
+    + `mstats_race`: dataset describing above information for all 2021-2022 matriculants by race and ethnicity in tidy format.
+    + `astats_sex`: dataset describing above information for all 2021-2022 applicants by sex in tidy format.
+    + `mstats_sex`: dataset describing above information for all 2021-2022 matriculants by sex in tidy format.
+    
+
+# medapps2021 0.0.0.9000
 
 * Added 2 cleaned datasets:
     + `astats_all`: dataset describing mean total MCAT exam/sub-section scores and combined/science/non-science GPAs for all 2021-2022 applicants in tidy format.